requirements.txt
langgraph
langchain
langchain-openai
openai

pandas
matplotlib
requests
python-dotenv

jupyter
